fix: combined user APIs to one file
