+++ src/app/api/users/route.ts
@@ -1,13 +1,19 @@
+import clientPromise from "@/app/lib/mongoDB";
 import connectToDB from "@/app/lib/mongoose";
 import Users from "@/app/models/user.model";
+import { getServerSession } from "next-auth";
 import { NextRequest, NextResponse } from "next/server";
+import { authOptions } from "../auth/[...nextauth]/options";
+import bcrypt from "bcrypt";
 
 export const dynamic = "force-dynamic";
 
 // to get users
 export async function GET(req: NextRequest) {
     try {
-        await connectToDB();
+        const client = await clientPromise;
+        const db = client.db();
+        // await connectToDB();
         const user_id = req.nextUrl.searchParams.get("user_id");
         const offset = Number(req.nextUrl.searchParams.get("offset")) || 0;
         const limit = Number(req.nextUrl.searchParams.get("limit"));
@@ -28,9 +34,12 @@ export async function GET(req: NextRequest) {
         }
 
         // api/users to get all users that are active
-        const users = await Users.find({ isActive: true })
+        const users = await db
+            .collection("users")
+            .find({ isActive: true })
             .limit(limit)
-            .skip(offset);
+            .skip(offset)
+            .toArray();
 
         if (users) {
             return NextResponse.json(
@@ -41,6 +50,105 @@ export async function GET(req: NextRequest) {
             return NextResponse.json(
                 { message: "No users found" },
                 { status: 404 }
+            );
+        }
+    } catch (error) {
+        console.error(error);
+        return NextResponse.json({ message: "Internal server error" });
+    }
+}
+
+// FIXME: to create user URL: /api/users (NOT WORKING YET)
+export async function POST(req: NextRequest) {
+    const session = await getServerSession(authOptions);
+    if (session?.user.role !== "owner" && session?.user.role !== "admin") {
+        return NextResponse.json(
+            {
+                message:
+                    "Unauthorized! Your role does not have access to this function",
+            },
+            { status: 400 }
+        );
+    }
+    try {
+        // await connectToDB();
+        const client = await clientPromise;
+        const db = client.db();
+        const requestBody = await req.json();
+        const createData: { [key: string]: any } = {};
+        if (requestBody) {
+            Object.keys(requestBody).forEach((key) => {
+                createData[key] = requestBody[key] as string | boolean | number;
+            });
+        }
+
+        const salt = await bcrypt.genSalt(10);
+        const hashedPassword = await bcrypt.hash(createData.password, salt);
+        const user = await db.collection("users").insertOne(createData);
+
+        if (user) {
+            return NextResponse.json(
+                { message: "user has been created" },
+                { status: 200 }
+            );
+        } else {
+            return NextResponse.json(
+                { message: "Cannot create user" },
+                { status: 404 }
+            );
+        }
+    } catch (error) {
+        console.error(error);
+        return NextResponse.json({ message: "Internal server error" });
+    }
+}
+
+// to edit user URL: /api/users/edit?user_id=657ab7edd422075ea7871f65
+export async function PUT(req: NextRequest) {
+    const session = await getServerSession(authOptions);
+    if (session?.user.role !== "owner" && session?.user.role !== "admin") {
+        return NextResponse.json(
+            {
+                message:
+                    "Unauthorized! Your role does not have access to this function",
+            },
+            { status: 400 }
+        );
+    }
+
+    console.log("User is Authorized!");
+
+    try {
+        await connectToDB();
+        const user_id = req.nextUrl.searchParams.get("user_id");
+
+        const requestBody = await req.json();
+        const editData: { [key: string]: string | boolean | number } = {};
+        if (requestBody) {
+            Object.keys(requestBody).forEach((key) => {
+                editData[key] = requestBody[key] as string | boolean | number;
+            });
+        }
+
+        if (user_id) {
+            const user = await Users.findOneAndUpdate(
+                { _id: user_id, isActive: true },
+                editData,
+                { new: true }
+            ).select("-password");
+
+            if (user) {
+                return NextResponse.json(user, { status: 200 });
+            } else {
+                return NextResponse.json(
+                    { message: "Cannot find User" },
+                    { status: 404 }
+                );
+            }
+        } else {
+            return NextResponse.json(
+                { message: "No ID has been provided" },
+                { status: 400 }
             );
         }
     } catch (error) {
